Add support for URI fragments with JSON Pointer in CI scripts
And use a single implementation of resolve_references for all scripts.
